Improve builder #225
