+++ app/src/test/java/ch/epfl/sweng/zuluzulu/EventTest.java
@@ -63,7 +63,7 @@ public class EventTest {
         map.put("category", CATEGORY);
         map.put("url_place_and_room", URL_PLACE);
 
-        event0 = new Event(new FirebaseMapDecorator(map));
+        event0 = new Event.EventBuilder().build(new FirebaseMapDecorator(map));
     }
 
     private void initDefaultAssociation() {
@@ -82,7 +82,7 @@ public class EventTest {
         map.put("category", CATEGORY);
         map.put("url_place_and_room", URL_PLACE);
 
-        event1 = new Event(new FirebaseMapDecorator(map));
+        event1 = new Event.EventBuilder().build(new FirebaseMapDecorator(map));
     }
 
     @Test(expected = IllegalArgumentException.class)
@@ -91,7 +91,7 @@ public class EventTest {
         map.put("id", 1L);
 
         FirebaseMapDecorator fmap = new FirebaseMapDecorator(map);
-        new Event(fmap);
+        new Event.EventBuilder().build(fmap);
     }
 
     @Test
@@ -150,28 +150,28 @@ public class EventTest {
     }
 
     @Test
-    public void dateComparTest () {
+    public void dateComparTest() {
         initWorkingAssociation();
         initDefaultAssociation();
         assertEquals(START_DATE.compareTo(START_DATE_1), Event.dateComparator().compare(event0, event1));
     }
 
     @Test
-    public void likesComparTest () {
+    public void likesComparTest() {
         initWorkingAssociation();
         initDefaultAssociation();
         assertEquals(-1, Event.likeComparator().compare(event0, event1));
     }
 
     @Test
-    public void increaseLikesTest () {
+    public void increaseLikesTest() {
         initWorkingAssociation();
         event0.increaseLikes();
         assertEquals(((int) LIKES_1) + 1, (int) event0.getLikes());
     }
 
     @Test
-    public void decreaseLikesTest () {
+    public void decreaseLikesTest() {
         initWorkingAssociation();
         event0.decreaseLikes();
         assertEquals(((int) LIKES_1) - 1, (int) event0.getLikes());
@@ -194,21 +194,25 @@ public class EventTest {
         initWorkingAssociation();
         assertEquals(SPEAKER, event0.getSpeaker());
     }
+
     @Test
     public void categoryIsCorrect() {
         initWorkingAssociation();
         assertEquals(CATEGORY, event0.getCategory());
     }
+
     @Test
     public void contactIsCorrect() {
         initWorkingAssociation();
         assertEquals(CONTACT, event0.getContact());
     }
+
     @Test
     public void websiteIsCorrect() {
         initWorkingAssociation();
         assertEquals(WEBSITE, event0.getWebsite());
     }
+
     @Test
     public void placeUrlIsCorrect() {
         initWorkingAssociation();
@@ -222,20 +226,20 @@ public class EventTest {
                 .setDate(new EventDate(
                         START_DATE, START_DATE))
                 .setUrlPlaceAndRoom(URL_PLACE)
-                .setLikes(      0)
-                .setShortDesc(  SHORT_DESC)
-                .setLongDesc(   LONG_DESC)
-                .setOrganizer(  ORGANIZER)
-                .setPlace(      PLACE)
-                .setBannerUri(  "epfl.ch")
-                .setIconUri(    "epfl.ch")
-                .setWebsite(    WEBSITE)
-                .setContact(    CONTACT)
-                .setName(       NAME1)
-                .setCategory(   CATEGORY)
-                .setSpeaker(    SPEAKER)
-                .setChannelId(  0)
-                .setAssosId(    0);
+                .setLikes(0)
+                .setShortDesc(SHORT_DESC)
+                .setLongDesc(LONG_DESC)
+                .setOrganizer(ORGANIZER)
+                .setPlace(PLACE)
+                .setBannerUri("epfl.ch")
+                .setIconUri("epfl.ch")
+                .setWebsite(WEBSITE)
+                .setContact(CONTACT)
+                .setName(NAME1)
+                .setCategory(CATEGORY)
+                .setSpeaker(SPEAKER)
+                .setChannelId(0)
+                .setAssosId(0);
 
         Event event0 = eb.build();
